requirements.txt
pynput
pyyaml
opencv-python
pybullet
click
scipy
numpy